refact: removing uncessary api bolerplate routes and keeping page component lean
