+++ apps/web/src/app/_components/header.tsx
@@ -1,0 +1,8 @@
+
+export function Header() {
+    return <header className="space-y-2">
+        <h1 className="text-3xl font-semibold tracking-tight text-foreground">
+            Flavor Map
+        </h1>
+    </header>;
+}
