+++ apps/web/src/app/_components/index.ts
@@ -1,0 +1,2 @@
+export * from "./header";
+export * from "./search-inbox";
