+++ apps/web/src/app/page.tsx
@@ -1,9 +1,12 @@
 import {
-  Button,
   Card,
   CardContent,
-  Input,
 } from 'ui-common';
+import { 
+  Header, 
+  SearchInbox
+} from './_components';
+
 
 export default function Index() {
   return (
@@ -12,20 +15,9 @@ export default function Index() {
 
       <Card className="border border-border/70 sm:max-w-sm md:max-w-md lg:max-w-lg bg-card">
         <CardContent className="flex flex-col gap-3 sm:flex-row sm:items-center">
-          <Input
-            placeholder="Search neighborhoods, cuisines, or places"
-            className="w-full"
-          />
-          <Button>Search</Button>
+          <SearchInbox />
         </CardContent>
       </Card>
     </main>
   );
-}
-export function Header() {
-    return <header className="space-y-2">
-        <h1 className="text-3xl font-semibold tracking-tight text-foreground">
-            Flavor Map
-        </h1>
-    </header>;
 }
